tiantian/requirements.txt
amqp==2.5.2
backcall==0.1.0
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
Django==2.2.3
django-haystack==2.8.1
django-redis==4.11.0
django-tinymce==2.8.0
fdfs-client-py==1.2.6
idna==2.8
importlib-metadata==1.4.0
ipython==7.11.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.2
jieba==0.42.1
kombu==4.6.7
more-itertools==8.1.0
mutagen==1.43.0
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.2
ptyprocess==0.6.0
Pygments==2.5.2
PyMySQL==0.9.3
pytz==2019.1
redis==3.3.11
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.8
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.1.8
Whoosh==2.7.4
zipp==2.0.1
